Fix the ABORT-INTERRUPT mess
Previously HW interrput handling was not correct. It should only invoke
interrupt handling when block does not get aborted. Fix this and also fix
HALT being aborted. Note that block failing due to something unrelated
is not an abort.

This syncs even in presence of all three of INSTRUCTION_START, HALT and
complete block aborts.