Fix emulation bug with self-modifying code
This patch adds an option FLUSHONMODIFY (defaulting to true for newly assembled
recordings) that causes the emulator to automatically invalidate caches of code
that modifies itself, even while running. This is necessary to get calls to
__int86 under DJGPP to emulate correctly, among other things; note that the
pipelining behaviour might not be exactly the same as on all models of x86
chips, though.